pyproject.toml
[tool.poetry]
name = "coworking-booking-backend"
version = "0.1.0"
description = "Backend for the coworking booking system."
authors = ["ITMO Project Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = ">=0.115,<1.0"
uvicorn = { version = ">=0.30,<1.0", extras = ["standard"] }
pydantic-settings = ">=2.6,<3.0"
sqlalchemy = ">=2.0,<3.0"
alembic = ">=1.14,<2.0"
asyncpg = ">=0.30,<1.0"
geoalchemy2 = ">=0.16,<1.0"
psycopg = { version = ">=3.2,<4.0", extras = ["binary"] }
PyJWT = ">=2.10,<3.0"
pwdlib = { version = ">=0.2,<1.0", extras = ["argon2"] }
email-validator = ">=2.2,<3.0"
greenlet = ">=3.1,<4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3,<9.0"
pytest-asyncio = ">=0.24,<1.0"
httpx = ">=0.28,<1.0"
ruff = ">=0.8,<1.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
